+++ src/Page.jsx
@@ -1,5 +1,5 @@
-import Input from './Input';
-import List from './List';
+import Input from './components/Input';
+import List from './components/List';
 
 export default function Page({
   taskTitle, onChangeTitle, onClickAddTask,
